笔记点赞、取消日增量变更数据入库

+++ xiaohongshu-data-align/src/main/java/com/zry/xiaohongshu/data/align/constant/TableConstants.java
@@ -12,7 +12,7 @@ public class TableConstants {
      * @param hashKey
      * @return
      */
-    public static String buildTableNameSuffix(String date, int hashKey) {
+    public static String buildTableNameSuffix(String date, long hashKey) {
         // 拼接完整的表名
         return date + TABLE_NAME_SEPARATE + hashKey;
     }
